Now with simple loop walking!
As far as I can see, the simple loop substitution for multiplication
operations is now functional.

To make this happen, I had to compare the IL that I was emitting with a
program containg the same C# instructions but compiled with the C#
compiler, and making sure that they are the same...and if they're not, I
would replicate the IL taken from the C# compiler.