Yet even more progress on static code analysis
The code walking is still a bit broken at the moment, but I'm currently
working on it.

The Walk method returns the final state of the Domain after the walk.
But since now it's walking over Write operations, I need to know more
than the final state of the domain because I need to take into account
the Writes in between the operations.
as well.