Updated the README file
I updated it mostly to add the following section:

NOTE: All of the below optimizations are heavily outdated.
Since the time when they were written, I've completely rearchitectured
the optimizition process with the introduction of the DIL (Dreas
Intermediate Language).

When I'll have a good working state of the compiler, I'll rewrite all of
the documentation for the new optimization process.